Cylinders work
Cylinders functional; both intersections and normals seem to work. There is perhaps an issue with the normals, as the stickman is very weirdly shaded (he has a lot of strange shadows). Also, check the edges of the cylinders, because I'm not completely convinced.
Stickman is now cylinders, but he is weirdly shaded