Fixed cylinder bug
Fixed small cylinder bug where we weren't transforming to local point on the cylinder. Cylinder person looks a lot better. Still very weird, but I think that's because of the hard shadows (the legs are like 4 pixels wide, but it does seem like there's some variety in the red, so some attempt at proper shading). But come to think of it, I'm surprised the stickperson looked anything close to normal when I had forgotten to convert the points.