Implemented Distributed RayTracing
Distributed raytracing with stratified sampling works for any number of bins. Might want to check that the numbers actually look right (i.e. the pixel positions in the bin), but the result with 36 bins looks way better (and takes a minute to render a frame...)
Non-distributed no longer works, since there'll be randomness regardless. We might want to add that back in at some point?